+++ Semana_4/node/src/@types/express.d.ts
@@ -1,0 +1,9 @@
+declare namespace Express {
+  // Sobreescrever uma tipagem do express
+  export interface Request {
+    // Anexa ao objeto Request
+    user: {
+      id: string;
+    };
+  }
+}