+++ Semana_4/node/src/config/auth.ts
@@ -1,0 +1,6 @@
+export default {
+    jwt:{
+        secret: "ea71a5c42d036b1448181fac1094c840",
+        expiresIn: "30m"
+    }
+}