+++ Semana_4/node/src/modules/appointments/repositories/IAppointmentsRepository.ts
@@ -1,0 +1,10 @@
+import Appointments from "@appointments/infra/Typeorm/entities/Appointments";
+import ICreateAppointmentsDTO from '@appointments/Dtos/ICreateAppointmentsDTO'
+
+
+interface IAppointmentsRepository {
+  create(CreateDTO:ICreateAppointmentsDTO): Promise<Appointments>;
+  findByDate(date: Date): Promise<Appointments | undefined>;
+}
+
+export default IAppointmentsRepository;